test(cli): split hosted marker contracts
